Try using the GH pages hosted version instead of self running

+++ convert.js
@@ -5,10 +5,17 @@ const puppeteer = require('puppeteer');
 const { spawnSync, spawn } = require('child_process');
 
 (async () => {
-	const jekyll = spawn("bundle", ["exec", "jekyll", "serve"]);
-
-	// Wait 5s for jekyll to be ready
-	await new Promise(resolve => setTimeout(resolve, 5000));
+	let local = process.argv.includes("local");
+	let jekyll = null;
+	if(local) {
+		console.log("Generating images from local files...");
+		jekyll = spawn("bundle", ["exec", "jekyll", "serve"]);
+		
+		// Wait 5s for jekyll to be ready
+		await new Promise(resolve => setTimeout(resolve, 5000));
+	} else {
+		console.log("Generating images from https://manual.ds-homebrew.com...");
+	}
 
 	const browser = await puppeteer.launch();
 	const tab = await browser.newPage();
@@ -24,7 +31,7 @@ const { spawnSync, spawn } = require('child_process');
 			let outPath = `nitrofiles/pages/${dir.substr(1)}/${page.substr(0, page.indexOf("."))}.gif`;
 			if(!fs.existsSync(outPath) || fs.statSync(outPath).mtime < fs.statSync(`pages/${dir}/${page}`).mtime) {
 				console.log(dir, page);
-				await tab.goto(`http://127.0.0.1:4000/${dir.substr(1)}/${page.substr(0, page.indexOf("."))}`, {waitUntil: "networkidle0"});
+				await tab.goto(`${local ? "http://127.0.0.1:4000/" : "https://manual.ds-homebrew.com/"}${dir.substr(1)}/${page.substr(0, page.indexOf("."))}`, {waitUntil: "networkidle0"});
 				await tab.screenshot({path: "screenshot.png"});
 				spawnSync("ffmpeg", ["-i", "screenshot.png", "-vf", "palettegen=max_colors=246", "palette.png", "-y"]);
 				const crop = spawnSync("ffmpeg", ["-loop", "1", "-i", "screenshot.png", "-frames:v", "3", "-vf", "negate,cropdetect=0:2:0", "-f", "null", "-"]).stderr.toString().match(/crop=.*:.*:.*:.*/)[0];
@@ -33,7 +40,7 @@ const { spawnSync, spawn } = require('child_process');
 
 			outPath = `nitrofiles/pages/${dir.substr(1)}/${page.substr(0, page.indexOf("."))}.ini`;
 			if(!fs.existsSync(outPath) || fs.statSync(outPath).mtime < fs.statSync(`pages/${dir}/${page}`).mtime) {
-				await tab.goto(`http://127.0.0.1:4000/${dir.substr(1)}/${page.substr(0, page.indexOf("."))}`);
+				await tab.goto(`${local ? "http://127.0.0.1:4000/" : "https://manual.ds-homebrew.com/"}${dir.substr(1)}/${page.substr(0, page.indexOf("."))}`);
 
 				let links = await tab.evaluate(() => {
 					let out = `
@@ -65,7 +72,8 @@ DEST = ${links[i].href.substr(links[i].href.lastIndexOf("/") + 1)}
 
 	await browser.close();
 
-	process.kill(jekyll.pid);
+	if(process.argv.includes("local"))
+		process.kill(jekyll.pid);
 
 	if(fs.existsSync("screenshot.png"))
 		fs.rmSync("screenshot.png");
